fix parse bug

+++ api/repositories/TextRepository.js
@@ -32,7 +32,7 @@ class TextRepository {
             uncertainty_ignore = parseFloat(result_ignore.rows[0].uncertainty);
           }
           
-          const uncertainty_threshold = parseInt(keys.uncertaintyThreshold);
+          const uncertainty_threshold = parseFloat(keys.uncertaintyThreshold);
           if (uncertainty_ignore) {
             if (uncertainty_ignore - uncertainty >= uncertainty_threshold) {
               return result_ignore.rows[0];
